+++ src/components/App.test.js
@@ -8,7 +8,7 @@ import App from './App';
 describe('App', () => {
   const mockStore = configureStore();
 
-  it('should render the intro without data, render the footer, and not the menu', () => {
+  it('should render a loader', () => {
     const store = mockStore({
       app: {
         storage: {
@@ -27,33 +27,7 @@ describe('App', () => {
         </Provider>
       </MemoryRouter>
     );
-    expect(container.find('Menu').length).toEqual(0);
     const rendered = container.render();
-    expect(rendered.find('h1').text()).toEqual('Off The Books');
-    expect(rendered.find('a').last().text()).toEqual('Source Code');
-  });
-
-  it('should render the intro with data', () => {
-    const store = mockStore({
-      app: {
-        storage: {
-          localStorage: false
-        }
-      },
-      transactions: {
-        data: [{}]
-      },
-    });
-
-    const container = shallow(
-      <MemoryRouter initialEntries={['/']}>
-        <Provider store={store}>
-          <App />
-        </Provider>
-      </MemoryRouter>
-    );
-    const rendered = container.render();
-    expect(rendered.find('h1').text()).toEqual('Welcome back!');
-    expect(rendered.find('p').first().text()).toEqual('You have 1 transactions so far.');
+    expect(rendered.find('div[title="Loading..."]').length).toEqual(1);
   });
 });
